+++ uPaths.pas
@@ -28,6 +28,9 @@ function AppDataDir: string;
 function IsPortable: Boolean;
 function ConfigFile: string;
 function DraftFile: string;
+{ Where the example drawings are put: beside the program when it is portable,
+  and in the user's folder when it is not. }
+function ExamplesDir: string;
 
 implementation
 
@@ -90,6 +93,11 @@ end;
 function DraftFile: string;
 begin
   Result := AppDataDir + 'heckers-sketch-draft.hsk';
+end;
+
+function ExamplesDir: string;
+begin
+  Result := AppDataDir + 'examples' + PathDelim;
 end;
 
 end.